backend-evaluator/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
load-dotenv==0.1.0
MarkupSafe==3.0.2
numpy==2.3.1
python-dotenv==1.1.1
scipy==1.16.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
pydantic==2.8.2
alembic==1.13.2
jiwer==3.0.4
requests==2.32.3
tqdm==4.66.5
python-dateutil==2.9.0.post0
matplotlib 
pandas